Converting a couple more functional tests to "fast"
Also starting to convert duplicated code in ConnectedRESTQA to use ml-app-deployer
